# backend/nodejs/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Beauty Salon Management API
  description: |
    API completa para gerenciamento de salão de beleza, incluindo clientes, serviços, funcionários e agendamentos.
    
    ## Funcionalidades
    - Gerenciamento de clientes
    - Catálogo de serviços
    - Gestão de funcionários
    - Sistema de agendamentos
    - Monitoramento e métricas
    
    ## Autenticação
    Esta versão da API não implementa autenticação. Em produção, considere implementar JWT ou OAuth2.
    
    ## Rate Limiting
    A API implementa rate limiting de 100 requisições por 15 minutos por IP.
  version: 1.0.0
  contact:
    name: Beauty Salon Team
    email: dev@beautysalon.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.beautysalon.com
    description: Servidor de produção
...
paths:
...
  /api/appointments:
    get:
      tags: [appointments]
      summary: Listar todos os agendamentos
      operationId: getAppointments
      responses:
        '200':
          description: Lista de agendamentos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [appointments]
      summary: Criar novo agendamento
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentInput'
      responses:
        '201':
          description: Agendamento criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/appointments/date/{date}/staff/{staffId}:
    get:
      tags: [appointments]
      summary: Listar agendamentos por data e funcionário
      description: Retorna uma lista de agendamentos filtrados por data e funcionário
      operationId: getAppointmentsByDateAndStaff
      parameters:
        - name: date
          in: path
          required: true
          description: Data dos agendamentos (formato YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2025-01-15"
        - name: staffId
          in: path
          required: true
          description: ID do funcionário
          schema:
            type: string
            format: uuid
            example: "a4e0423e-8243-4ff4-9341-9875f6b4a2d8"
      responses:
        '200':
          description: Lista de agendamentos da data e funcionário retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/appointments/{id}:
    get:
      tags: [appointments]
      summary: Buscar agendamento por ID
      operationId: getAppointmentById
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Agendamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags: [appointments]
      summary: Atualizar agendamento
      operationId: updateAppointment
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentInput'
      responses:
        '200':
          description: Agendamento atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags: [appointments]
      summary: Excluir agendamento
      operationId: deleteAppointment
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Agendamento excluído com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/staff/{id}:
    get:
      tags: [staff]
      summary: Buscar funcionário por ID
      description: Retorna os dados de um funcionário específico
      operationId: getStaffById
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Funcionário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags: [staff]
      summary: Atualizar funcionário
      description: Atualiza os dados de um funcionário existente
      operationId: updateStaff
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffInput'
      responses:
        '200':
          description: Funcionário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags: [staff]
      summary: Excluir funcionário
      description: Remove um funcionário do sistema
      operationId: deleteStaff
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Funcionário excluído com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: ID único do recurso
      schema:
        type: string
        format: uuid
  schemas:
...
    Staff:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Ana Souza"
        role:
          type: string
          example: "Hairstylist"
        specialties:
          type: array
          items:
            type: string
          example: ["Coloração", "Corte a seco"]
        phone:
          type: string
          example: "11999998888"
        email:
          type: string
          format: email
          example: "ana.souza@beautysalon.com"
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Appointment:
      type: object
      required: [id, customerId, serviceId, staffId, appointmentTime]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        appointmentTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, completed, cancelled]
          default: scheduled
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [customerId, serviceId, staffId, appointmentTime]
      properties:
        customerId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        appointmentTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, completed, cancelled]
          default: scheduled
        notes:
          type: string
    StaffInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          example: "Ana Souza"
        role:
          type: string
          example: "Hairstylist"
        specialties:
          type: array
          items:
            type: string
          example: ["Coloração", "Corte a seco"]
        phone:
          type: string
          example: "11999998888"
        email:
          type: string
          format: email
          example: "ana.souza@beautysalon.com"
        isActive:
          type: boolean
          default: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Required field 'name' is missing"
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'